Work on sign screen proposal
